Preparing ECDSA for fhirlabs.net deploy
Also dropped FileCertifiateStore in favor of smaller alternatives that load different certificates base on whether the code is a client or server.
