+++ examples/FhirLabsApi/appsettings.Production.json
@@ -89,6 +89,17 @@
           "RegistrationEndpoint": "https://securedcontrols.net/connect/register"
         }
       }
+      {
+        "Community": "udap://ECDSA/",
+        "SignedMetadataConfig": {
+          "SigningAlgorithm": "ES384",
+          "Issuer": "https://fhirlabs.net/fhir/r4",
+          "Subject": "https://fhirlabs.net/fhir/r4",
+          "AuthorizationEndPoint": "https://securedcontrols.net/connect/authorize",
+          "TokenEndpoint": "https://securedcontrols.net/connect/token",
+          "RegistrationEndpoint": "https://securedcontrols.net/connect/register"
+        }
+      }
     ]
   }
 }
